refactor(glossary): move to client folder
