+++ site/client/GlossaryExcerpt/GlossaryExcerpt.tsx
