+++ site/client/owid.scss
@@ -55,7 +55,7 @@
 @import "site/client/blocks/Summary/summary.scss";
 @import "site/client/blocks/Help/help.scss";
 @import "site/client/blocks/CookiePreferences/CookiePreferences.scss";
-@import "site/GlossaryExcerpt/GlossaryExcerpt.scss";
+@import "site/client/GlossaryExcerpt/GlossaryExcerpt.scss";
 
 /* HACK (Mispy): Fix search autozooming on iPhone Safari.
    http://stackoverflow.com/questions/2989263/disable-auto-zoom-in-input-text-tag-safari-on-iphone */
